+++ doc/source/changelog.rst
@@ -7,6 +7,8 @@ next
 * Give preference to None parsers in Unions.
 * `defopt.signature` annotates the return type with the documented raisable
   exception types.
+* Added `defopt.bind` to allow preprocessing arguments before performing the
+  call.
 
 6.1.0 (2021-02-25)
 ------------------