+++ doc/source/changelog.rst
@@ -3,6 +3,7 @@ Changelog
 
 next
 ----
+* Added support for Python 3.10.
 * Fixed bad interaction between custom None parsers and tuple parsers.
 * Give preference to None parsers in Unions.